Confirm dialog

+++ src/app/cart/cart.component.ts
@@ -1,6 +1,9 @@
 import { Component, EventEmitter, OnInit, Output } from '@angular/core';
 import { CartService } from '../services/cart.service';
 import {ArtPiece} from "../models/artPiece.model";
+import {ConfirmDialogComponent} from "../confirm-dialog/confirm-dialog.component";
+import {MatDialog} from "@angular/material/dialog";
+import {NgbModal} from "@ng-bootstrap/ng-bootstrap";
 
 @Component({
   selector: 'app-cart',
@@ -19,7 +22,11 @@ export class CartComponent implements OnInit {
   itemCount: Number;
   inCart: boolean;
 
-  constructor(  private _cartService: CartService) { }
+  constructor(
+    private _cartService: CartService,
+    private dialog: MatDialog,
+    private modalService: NgbModal
+  ) { }
 
   ngOnInit(): void {
     this.calculateTotal()
@@ -58,13 +65,39 @@ export class CartComponent implements OnInit {
   }
 
   removeItem(item){
-    this._cartService.removeFromCart(item);
-    this.itemCount= this._cartService.getCount();
-    console.log("count: ", this.itemCount)
-    this.cartCount.emit(this.itemCount);
-    this.inCart = false;
-    this.removeLocally(item);
-    this.totalPrice = this.reCalculateTotal();
-    (<ArtPiece>item).inCart = false;
+
+    const modalRef = this.modalService.open(ConfirmDialogComponent)
+    modalRef.componentInstance.title = 'Remove Item'
+    modalRef.componentInstance.message = 'Are you sure you want to remove this item from your cart?'
+    modalRef.result.then((result) => {
+      if (result) {
+        this._cartService.removeFromCart(item);
+        this.itemCount= this._cartService.getCount();
+        this.cartCount.emit(this.itemCount);
+        this.inCart = false;
+        this.removeLocally(item);
+        this.totalPrice = this.reCalculateTotal();
+        (<ArtPiece>item).inCart = false;
+      }
+    });
+
+    /*const confirmDialog = this.dialog.open(ConfirmDialogComponent, {
+      data: {
+        title: 'Remove Item',
+        message: 'Are you sure you want to remove this item from your cart?'
+      }
+    });
+    confirmDialog.afterClosed().subscribe(result => {
+      if (result === true) {
+        this._cartService.removeFromCart(item);
+        this.itemCount= this._cartService.getCount();
+        console.log("count: ", this.itemCount)
+        this.cartCount.emit(this.itemCount);
+        this.inCart = false;
+        this.removeLocally(item);
+        this.totalPrice = this.reCalculateTotal();
+        (<ArtPiece>item).inCart = false;
+      }
+    });*/
   }
 }
